Fix: Let connection pool handle connection
When creating a sqlalchemy session for a Context object, the old
implementation was binding the Session to an explicitly created database
connection. The Session object was not closing the connection in this
case, leading to a huge amount of open but unused connections on the
database. Now passing the engine object itself to the Session, leading
to the correct behaviour.