New: Support multiple transactions in a single ctx
The module now supports committing the transaction before destroying the
session. This feature is required because of a change in score.ctx:

https://github.com/score-framework/py.ctx/commit/6cfb6439c774dd1f9f02ceccf1c98aa9ca45de15

We are now instructing zope.sqlalchemy to retain the SA session after
committing the transaction and closing the session manually at the end
of the ctx lifetime.

This also introduces a breaking change: get_session() no longer works on
arbitrary Context objects. They must be bound to the same
ConfiguredCtxModule as the ConfiguredSaOrmModule.